fastApiApp/requirements.txt
annotated-types==0.7.0
annotated-types==0.7.0
anyio==4.8.0
anyio==4.8.0
cfgv==3.4.0
cfgv==3.4.0
click==8.1.8
click==8.1.8
distlib==0.3.9
distlib==0.3.9
Faker==35.2.0
Faker==35.2.0
fastapi==0.115.8
fastapi==0.115.8
filelock==3.17.0
filelock==3.17.0
h11==0.14.0
h11==0.14.0
identify==2.6.6
identify==2.6.6
idna==3.10
idna==3.10
nodeenv==1.9.1
nodeenv==1.9.1
platformdirs==4.3.6
platformdirs==4.3.6
pre_commit==4.1.0
pre_commit==4.1.0
psycopg2-binary==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic==2.10.6
pydantic_core==2.27.2
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-dotenv==1.0.1
PyYAML==6.0.2
PyYAML==6.0.2
six==1.17.0
six==1.17.0
sniffio==1.3.1
sniffio==1.3.1
SQLAlchemy==2.0.37
SQLAlchemy==2.0.37
starlette==0.45.3
starlette==0.45.3
typing_extensions==4.12.2
typing_extensions==4.12.2
uvicorn==0.34.0
uvicorn==0.34.0
virtualenv==20.29.1
virtualenv==20.29.1
